+++ src/main/java/kitchenpos/ordertable/domain/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.model;
+package kitchenpos.ordertable.model;
 
 public class OrderTable {
     private Long id;
